pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-pool-agent"
version = "0.1.0"
description = "A crypto pool agent project"
requires-python = ">=3.8"
dependencies = [
    "streamlit==1.40.0",
    "langchain==0.2.14",
    "langchain-community==0.2.12",
    "langchain-core==0.2.32",
    "langchain-experimental==0.0.64",
    "langchain-groq",
    "langchain-openai",
    "langchain-pinecone==0.1.3",
    "langgraph==0.2.14",
    "markdown-pdf==1.3",
    "matplotlib==3.9.2",
    "plotly==5.24.1",
    "plotly-express==0.4.1",
    "tabulate==0.9.0",
    "seaborn==0.13.2",
    "python-dotenv==1.0.1",
    "pandas==2.2.2",
    "pinecone==5.3.1",
    "pinecone-client==5.0.1",
    "pillow==10.4.0",
    "snowflake-connector-python==3.12.2",
    "snowflake==0.12.1",
    "kaleido",
    "streamlit-float",
    "streamlit-option-menu",
    "streamlit_js_eval==0.1.7",
    "streamlit-extras",
    "pydantic-ai==0.0.20",
    "logfire",
    "nest-asyncio",
    "gql"
]

[tool.hatch.build.targets.wheel]
packages = ["."] 